% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSignificatif.R
\name{TestSignificatif}
\alias{TestSignificatif}
\title{TestSignificatif}
\usage{
TestSignificatif(serie, p, q)
}
\arguments{
\item{serie}{série stationnaire (d=0)}

\item{p}{ordre partie autorégressive à tester}

\item{q}{ordre partie moyenne mobile à tester}
}
\value{
Significativité au seuil de 95\% des coefficients des ordres les plus élevés
}
\description{
TestSignificatif
}
