% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armafit.R
\name{armafit}
\alias{armafit}
\title{Title}
\usage{
armafit(serie, p, q)
}
\arguments{
\item{serie}{série stationnaire d'entrée}

\item{p}{ordre partie autorégressive à tester}

\item{q}{ordre partie moyenne mobile à tester}
}
\value{
TestAutocorr et TestSignificativite
}
\description{
Title
}
