% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armafit.R
\name{armafit}
\alias{armafit}
\title{Title}
\usage{
armafit(serie, p, q, lagmax)
}
\arguments{
\item{serie}{série Stationnaire d'entrée}

\item{p}{ordre partie autorégressive à tester}

\item{q}{ordre partie moyenne mobile à tester}

\item{lagmax}{lag max à tester pour les autocorrélations des résidus (statistique de Ljung-Box) > hypothèse forte résidus suivent un bruit blanc, test jusqu'à lag_max}
}
\value{
Synthese de TestAutocorr et TestSignificativite & Critères AIC BIC
}
\description{
Title
}
